// src/app/core/api/generated/model/payslipDto.ts
/**
 * Clean Architecture WebAPI
 * An example of Clean Architecture WebAPI with Swagger and JWT Authentication
 *
 * The version of the OpenAPI document: v1
 * Contact: kamrantajerbashi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PayslipDto { 
    id?: string;
    runId?: string;
    payslipNumber?: string | null;
    generatedAt?: string;
    status?: number;
}