// src/app/core/api/generated/model/payrollRunDto.ts
/**
 * Clean Architecture WebAPI
 * An example of Clean Architecture WebAPI with Swagger and JWT Authentication
 *
 * The version of the OpenAPI document: v1
 * Contact: kamrantajerbashi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PayrollRunDto { 
    id?: string;
    periodId?: string;
    employeeId?: string;
    baseSalary?: number;
    overtimeAmount?: number;
    absenceDeduction?: number;
    lateDeduction?: number;
    grossSalary?: number;
    totalDeductions?: number;
    netSalary?: number;
    status?: number;
    calculatedAt?: string;
}